Move fpm test program to subdirectory
Since each unit test is include'ed and not use'ed, there is no need to
compile each unit test. In fact, doing so, and linking all during the
test build, I believe, is what causes the Windows build failure, where
the command is "too long" for msys2 environment and causes a test
failure.